# api/openapi.yaml
openapi: 3.1.0
info:
  title: Sonic API
  version: 0.1.0
paths:
  /orders:
    post:
      summary: Create Order
      operationId: create_order_orders_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /positions/{position_id}/adjust:
    post:
      summary: Adjust Position
      operationId: adjust_position_positions__position_id__adjust_post
      parameters:
      - name: position_id
        in: path
        required: true
        schema:
          type: string
          title: Position Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionAdjustRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /alerts/liquidations:
    post:
      summary: Create Liquidation Alert
      operationId: create_liquidation_alert_alerts_liquidations_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: idempotency-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LiquidationAlertRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Alert:
      properties:
        id:
          type: string
          title: Id
        position_id:
          type: string
          title: Position Id
        level:
          type: string
          title: Level
          default: warning
      type: object
      required:
      - id
      - position_id
      title: Alert
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LiquidationAlertRequest:
      properties:
        position_id:
          type: string
          title: Position Id
      type: object
      required:
      - position_id
      title: LiquidationAlertRequest
    Order:
      properties:
        id:
          type: string
          title: Id
        instrument:
          type: string
          title: Instrument
        side:
          type: string
          title: Side
        type:
          type: string
          title: Type
        qty:
          type: string
          title: Qty
        price:
          anyOf:
          - type: string
          - type: 'null'
          title: Price
        status:
          type: string
          title: Status
          default: open
        account_id:
          type: string
          title: Account Id
      type: object
      required:
      - id
      - instrument
      - side
      - type
      - qty
      - account_id
      title: Order
    OrderCreate:
      properties:
        instrument:
          type: string
          title: Instrument
        side:
          type: string
          title: Side
        type:
          type: string
          title: Type
        qty:
          type: string
          title: Qty
        price:
          anyOf:
          - type: string
          - type: 'null'
          title: Price
        account_id:
          type: string
          title: Account Id
        client_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Ref
      type: object
      required:
      - instrument
      - side
      - type
      - qty
      - account_id
      title: OrderCreate
    Position:
      properties:
        id:
          type: string
          title: Id
        instrument:
          type: string
          title: Instrument
        qty:
          type: string
          title: Qty
        account_id:
          type: string
          title: Account Id
      type: object
      required:
      - id
      - instrument
      - qty
      - account_id
      title: Position
    PositionAdjustRequest:
      properties:
        delta:
          type: integer
          title: Delta
      type: object
      required:
      - delta
      title: PositionAdjustRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError